+++ CHANGELOG.rst
@@ -7,6 +7,7 @@ v1.2.5
 *In development*
 
 - Fix firefox initialization error using Selenium 2.X
+- Add *wait_until_loaded* method to PageObject class to wait until all page elements with wait=True are visible
 
 v1.2.4
 ------